requirements.txt
python-telegram-bot==20.7
python-binance
python-dotenv
ta
tqdm
scikit-learn
matplotlib
keras
tensorflow
seaborn
xgboost
tabulate
tree